Refactor how visualisations modify the code.
The webview can now notify a particular visualisation to indicate that the code should be modified in a certain way. The visualisation (extension side) is therefore left responsible for applying the modification.

This better separates the responsability between the view and the model of the visualisations. It also simplifies more advanced code transformations.