Enable visualisations to require a new parsing of the doc.
In this case, when new visualisations are generated by the core and sent to the webview, the message now carries a new flag which indicates whether it was requested by a visualisation or not.

This flag is meant to be used to decide whether a displayed visualisation should be updated (if it requested it) or not (e.g. if the code was changed in the editor, possibly breaking the old source index-to-vis. mapping in the webview).