Refactor the project structure and the build process.
This is the first step to generate the webview from modular visualisation (with separate view/model in TypeScript and static files to inject).

The new build process relies on separate tsconfig files + Rollup with a custom plugin to bundle the webview into a single HTML file (not in use for now).