Refactor and type the structure of the webview.
In particular, the PDF and the visualisations have two dedicated managers, which delegate most of the work to dedicated classes.

The WebviewManger class is responsible for instanciating both + a messenger (to communicate with the core).

This commit also set up boilerplate code for refactoring and typing the actual views of the three existing visualisations.