Re-implement the three existing views in TypeScript.
The view now receives some information about the PDF page which contains the annotation mask associated with the visualisation when it is instancited. This was required by the current implementation of the view of the includegraphics visualisation.

Note: the views have not been tested yet, and may need debugging before working again!