Ensure source files use normalised absolute paths.
This commit should fix some cases in which LaTeX wrote relative paths instead of absolute ones.

Nonetheless, since it is hard to be certain it will always be the case (e.g. it seems to happen when lualatex is used instead of pdflatex), relative paths in the mapping file are now considered relative to the directory containing the main document (instead of failing immediately).

In addition, to avoid mixing different kinds of paths pointing to the same resource, all paths from the mapping file are now normalised as well.