Fix/improve the Rollup plugin for creating inlined HTML
The new plugin fix an issue due to the special meaning of certain special expressions in String.replace replacement strings (in that case: $').

It also add an option to minify the output HTML—but it is disabled by default (as it takes ~25s to output minified HTML instead of ~3s on my computer). It could be better configured and used for production builds.