Restrict when the visualisations can be used
This commit implements several new features (e.g. saving all source files at once, testing if some source file is dirty, etc) as side additions to implement the main new feature: preventing users from editing the code from a visualisation (in the webview) after a file with visualisable code has been modified but not saved, and therefore not recompiled (resulting in displacements between the old positions known by the models and hte new positions in the document).

This change is quite restrictive (and probably requires some fixes, e.g. when a visualisation is opened while the webview must disallow their usage), but it ensures a safe usage of visualisations.

A better alternative would be to track when changes occur INSIDE visualisable pieces of code, so that re-parsing the document and re-building the visualisation models could be sufficient. However, this requires to keep track of changes to adapt the lines in the code mappings (e.g. in case the user inserts new lines somewhere in the document, shifting the line numbers of everything below and therefore preventing iLaTeX from finding the pieces of code at the line numbers written by LaTeX).