Refactor the AST visitor pattern
The main ASTVisitor interface now require one method per type of node; each AST node is responsible for calling the right method on itself.

The ASTVisitorAdapter class is now only responsible for making each specific visit method call a default (empty, overridable) method. It is not responsible for selecting the right method to call depending on the type of the node anymore!