Refactor the LaTeX parsers into dedicated classes
The new 'LatexParser' class is now responsible for parsing a source file into an AST (or failing properly).

This change, among other things, enable to parameterise the parsers with values at runtime (such as the source file associated with a particular parsing operation).

This, in turn, enables to provide more data to AST nodes when they are created. Tthey are now given the source file they have been parsed in + a 'reparser' function (so they can try to re-parse themselves if they are modified).

Note: all these changes are still local to the 'ast module' of the extension; the rest of the code has not been adapted yet.