Change the strategy used to update the content of one vis.
Previously, the update used to rely on the UID of the visualisation to match the old content node with the new content node. This approach was historically chosen to prevent issues when updating visualisation content in the webview while code mappings may have changed (e.g. if the document was recompiled while a popup was open).

This should/will not be a problem anymore, since the webview is notified every time the document is recompiled.

Therefore, in order to correctly update visualisation content, the webview now relies on the code mapping ID.