Refactor/re-implement the includegraphics vis.
Note: the view seems to display the image without trimming it correctly in certain cases.

This may be related to the fact the view does not handle content update for the moment, and/or the fact the AST does not handle reparsing for the moment. It should be fixed in a future commit.