Rewrite the AST range/node update system
Position shifts due to changes in a source file are now directly computed in the RangeInFile class (ranges must be given a SourceFileChange object, but the algorithm is otherwise the same). This commit also fixes a typo that caused some updated positions to have a negative column index.

The AST node update mechanism has been simplified. Since the only nodes whose updates matter are nodes used by a visualisation, nodes will NOT update themselves unless a special flag has been set (each visualisation model is responsible for that).

During a change processing, the new update algorithm works as follows:
– if a node is modified across its range, or within its range if it is a leaf, it is marked as "out of sync";
– if a node has at least one child node marked as "out of sync", it is marked as "out of sync" itself;
– if a node with the special flag on is marked "out of sync", it will try to reparse and replace itself (once the change has been processed by all its descendants).
– visualisation models can (must) observe node replacements ("updates") and update their own reference to the newly parsed AST node (if the reparsing was successful).

This is not working very well at the moment, at least because re-parsers of commands and environments must be fixed (as en example, reparsing an itabular environment worked, but it did not include any parameter, which made the visualisation model failed when it tried to read them). This should be fixed in future commits.