Refactor and improve visualisation view "contexts"
The new contexts can dynamically fetch the current PDF page detail/annotation mask that match with the ones that were used when the visualisation view was instanciated.

For example, this can be used to retrieve the current width/height/scale of the PDF page that contain an open visualisation view, whose update depends on these information.