Change the grid layout env. and visualisation
The new grid layout system requires to specify the total width and height of the grid, and it makes the required row heights relative instead of absolute. The parser, the model and the view have been adapted to take this change into account.

The communication between the view and the model has not been fully reimplemented. It will be done in another commit.