Rewrite the source file edit system and improve/fix vis. edit operations
The new system is based on the 'AtomicSourceFileEditor' class, that provides the features required to perform edits (in a similar fashion than what VS Code provides, but with the ability to accumulate async edit providers that will provide actual edit operations when they are resolved at edit time).

Using the system, the edit operations on AST nodes, the layout of the grid layout vis. and the grid of the tabular vis. have been moved in dedicated files (as edit providers), while the base classes are responsible for creating editors, passing edits to the editor, and applying them.

Several issues with the edit operations of the grid layout vis. have normally been fixed during the transition.