Add an demo version of the IHM'20'21 paper on iLaTeX
Note: the metadata of all PNG images had to be removed from the files to fix a weird bug that broke the way the \includegrahics command trims images. The bug seems specific to PNG images, as it did not happen when they were converted to JPEG images. This could require a fix in the future (e.g. to take the pixel density metadata into account in the visualisation, as it seems to have a role here).