Add an iLaTeX mode without any visualisations
This mode prevents iLaTeX from reading code mappings, creating/parsing source files, creating visualisations, and displaying them in the webview.

The "ilatex.init" command has been renamed to "ilatex.initWithVisualisations", and a new "ilatex.initWithoutVisualisations" is now provided.

This "global option" is passed to the InteractiveLaTeX class constructor, stored in a dedicated property in every iLaTeX instance, and a new type of message enables the core to send new global options to the webview.