Add an event logging + log file management system
The expected data of a log depends on the "source" of the log. Four sources are readily available (core, visualisations, user edits and errora).

Every log is stored in an internal array AND written to a special log file that is a CSV with a special header (one per main document). If a log file already exists, new logs are appended at the end; if not, it is created when the iLaTeX instance is created.