Refactor and improve the entry point of the extension
The extension.ts file now delegates every setup tasks to the InteractiveLatexExtensionContext singleton.

The latter is responsible for setting up a new InteractiveLatexDocumentManager (that creates, manages and destroys what used to be named "ilatex instances").

It also takes care of providing commands that are part of the extension and to create a new status bar item.

New commands are provided, among which a command that displays an interactive menu to create/display an iLaTeX document from either the active editor or a file. Variants to create an iLaTeX document without visualisations are provided. Clicking the status bar item will display this menu.

In order to give access to this menu and to display the status bar item even when there is no iLateX document, the extension is now started when VSCode starts.