Prevent giving focus to editor when vis. perform edits
This change enables to edit the document by manipulating visualisations without having to have the edited file displayed next to the PDF (before that, every edit would open the edited document in an editor with focus on every edit).

This commit changes a bit the SourceFile and the source file editors APIs. In particular, edits are now performed using WorkspaceEdit objects, and by default, lightweight editors now skip intermediate edits when the source file they operate on is not displayed in a visible editor.

This commit also fixes code selections in source files.