Improve how code mapping paths are handled
In case a code mapping path is relative, it is now considered relative to the main LaTeX file.
If a file at this new path exists, it is considered as the absolute path of this code mapping.
If not, a warning message is displayed and an exception is thrown.

When the parsing of a code mapping throw an exception, it is now caught by the code mapping manager, that will simply discard this code mapping.