Fix a bug introduced by a recent package update
In a recent commit, various packages were updated to more recent versions. Parsimmon, in particular, was bumped from 1.16.0 to 1.18.1.

However, starting from version 1.17, the semantic of getting the position of the parser's head after reading a newline character (\n) seems to have changed, as suggested by https://github.com/jneen/parsimmon/issues/331 (instead of returning a position at the end of line N where \n was read, it returns a position at the beginning of line N+1). This was a breaking change, given how i-LaTeX's parser is implemented. For this reason, the package was reverted to version 1.16.0.