Use a different method to determine if the LaTeX compilation failed
In some situations, relying on the error code latexmk terminates with seems brittle, as a code such as 12 does not necessarily indicate a compilation error, but simply a sort of warning.

To prevent treating the compilation as a failure when it actually succeeded, this commit switches to a less conservative approach: testing if the PDF file (still) exists and if the modification date is ulterior to the beginning of the compilation.

This can cause problems in situations in which, e.g., the file was indeed modified but contains erroneous code and cannot be parsed properly by i-LaTeX.